+++ src/building.cpp
@@ -2,10 +2,8 @@
 #include "..\\Logger\\logger.h"
 
 
-Building::Building(int n, int m, BuildType* type, size_t max)
+Building::Building(int n, int m, BuildType* type, unsigned int max)
 :loc(n, m), type(type), rent(type->calcRent(loc, 100, 100)), MaxCitCount(max), ppl() { AUTO_LOG(); }
-
-
 
 bool Building::addPerson(Citizen person)
 {
